update benchmark script

+++ java/fury-benchmark/analyze.py
@@ -60,7 +60,7 @@ def process_data(filepath: str):
 
 
 def plot(df: pd.DataFrame, file_dir, filename, column="Tps"):
-    df["ns"] = 1 / df["Tps"] * 10**9
+    df["ns"] = (1 / df["Tps"] * 10**9).astype(int)
     data = df.fillna("")
     data.to_csv(f"{file_dir}/pd_{filename}")
     if "objectType" in data.columns:
